Update supported-g-code.md
Added all known G and M Codes as found through a decompile of the M3D source. Also added some links to the standards mentioned in the page.  Please update any of the items marked as unknown or that mention possibly or probably. Some codes I was not able to definitely determine what they do from looking in the source alone.